small warning fix for powerpc-darwin
2006-05-24  Dalibor Topic  <robilad@kaffe.org>

        Small warning fix for powerpc-darwin.

        * include/jsyscall.h (_sendto) Modified prototype to match
        OS X & GNU/Linux declaration.

        * kaffe/kaffevm/systems/unix-jthreads/syscalls.c,
        kaffe/kaffevm/systems/unix-pthreads/syscalls.c
        (jthreadedSendto): Updated declaration to match libc.

+++ include/jsyscall.h
@@ -79,7 +79,7 @@ typedef struct SystemCallInterface {
 		int timeout, ssize_t *);
 	int	(*_sockwrite)(int, const void *, size_t, ssize_t *);
 	int	(*_sendto)(int, const void *, size_t, int, const struct sockaddr *,
-		socklen_t, int *);
+		socklen_t, ssize_t *);
 	int	(*_setsockopt)(int, int, int, const void *, int);
 	int	(*_getsockopt)(int, int, int, void *, socklen_t *);
 	int	(*_getsockname)(int, struct sockaddr *, socklen_t *);
